Jan 15th 1:07 am
Restarted from last commit to get back to a working file
Changed placement and order of photos.
Added a unit push and unit pull to the info and male photo to get the
arrangement to the specs given, what this does it let the medium and
small have a better feel to them have the info pushed to the bottom and
the photos at the top.